requirements.txt
click=6.6
Flask=0.11.1
Flask-SQLAlchemy=2.1
itsdangerous=0.24
Jinja2=2.8
MarkupSafe=0.23
PyYAML=3.12
requests=2.11.1
SQLAlchemy=1.0.15
Werkzeug=0.11.11